Reuse the sm90 im2col interface for conv2d/conv2d_dgrad (#181)
* Create new cutlass header files for xe4 workloads enabling

* Include xe4 files in common header files

* Update CMakeLists.txt

* Add xe4 inline pisa code 20250307

* Add xe4 bgemm test

* Add conv2d test

* fix error & add static check (#163)

* fix error

* add static_assert for static check

* fix: fix private mem in conv (unroll, inline, rm assert) (#164)

* rm assert and no_unroll

* inline to_underlying_arguments

* use bitwise and to reduce goto (#165)

* Fix copy traits/mma traits code (#167)

* Update README

* Remove useless file

* Apply sm90 PipelineTmaAsync in xe4 workloads (#169)

* Make sm90 pipeline work on xe4 platform

* Update gemm test

* Update pipeline usage in xe4 conv2d workload

* Add cuda arch check around brkpt asm

* Modify SYCL_INTEL_TARGET

* align the tensor copy coordinate definition with Nivida in conv2d and conv2d_dgrad

* merge the first 3 changes by shuangpe

* merge the put group infor into KernelTmaWarpSpecializedXe4 change

* merged Move cluster mask calculation to collective mma change

* merged Provide ClusterShape to xe4 tile scheduler change

* merged Remove PersistentTileSchedulerXe4::to_underlying_arguments change

* merged the Modify the interface of xe4 copy operations change

* merged Reuse sm90/sm100 make tiled copy code change

* merged Set tensor desc in device kernel code change

* merged Move fusion callback params to device side change

* remove XE4_IM2COL_COPY_Unpack

* first draft to merge copy_traits_xe4_im2col to copy_traits_sm90_im2col

* use nvidia Im2ColTmaDescriptor for xe4 row copy descriptor

* create tma_desc and tma_tensor together in make_im2col_tma_copy_desc

* second draft to merge copy_traits_xe4_im2col to copy_traits_sm90_im2col

* code clean up

* use make_tma_copy for matrix B in conv2d

* merge the 173 changes

* merge 174 change

* merged the change till 177

* merged change 178

* merged change 180

* remove DmaCache, namespace details and copy_traits_xe4_im2col.hpp

* fixed the issues in the comments

---------

+++ examples/sycl/xe4/conv2d/conv2d.cpp
@@ -60,10 +60,10 @@ int run_test(const conv2d::problem_shape_t &problem_shape)
     uint32_t dilation_h = problem_shape.get_dilation_h();
     uint32_t dilation_w = problem_shape.get_dilation_w();
 
-    using ElementAct = bf16;
-    using ElementFlt = bf16;
+    using ElementAct = fp16;
+    using ElementFlt = fp16;
     using ElementAcc = float;
-    using ElementOut = bf16;
+    using ElementOut = fp16;
 
     constexpr uint32_t Stages = 3;
     static constexpr auto tnspA = xe4::GMMA::Major::K;
